requirements.txt
PyQt5>=5.15
PyGObject>=3.46
numpy>=1.23
pyyaml>=6.0
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.1
opencv-python>=4.8.0
